修改绑定token到redis

+++ morning-run-biz/src/main/java/cn/doublehh/sport/service/impl/GradeServiceImpl.java
@@ -124,12 +124,14 @@ public class GradeServiceImpl extends ServiceImpl<GradeMapper, Grade> implements
                 grade.setUpdateTime(LocalDateTime.now());
                 grade.setSemesterId(semester);
                 //添加成绩到临时表
-                List<Grade> oldGradeTmpList = gradeMapper.getGradeTmpOld(grade);
-                if (!CollectionUtils.isEmpty(oldGradeTmpList)) {
-                    oldGradeTmpList.forEach(oldGradeTmp -> {
-                        oldGradeTmp.setIsValid(0);
-                        gradeMapper.deleteOldGradeTmp(oldGradeTmp);
-                    });
+                if (!"01".equals(grade.getType())) {
+                    List<Grade> oldGradeTmpList = gradeMapper.getGradeTmpOld(grade);
+                    if (!CollectionUtils.isEmpty(oldGradeTmpList)) {
+                        oldGradeTmpList.forEach(oldGradeTmp -> {
+                            oldGradeTmp.setIsValid(0);
+                            gradeMapper.deleteOldGradeTmp(oldGradeTmp);
+                        });
+                    }
                 }
                 gradeMapper.insertTmpGrade(grade);
             }
@@ -209,7 +211,7 @@ public class GradeServiceImpl extends ServiceImpl<GradeMapper, Grade> implements
                 templateMessage.addData(new WxMpTemplateData("keyword2", tsUser.getName(), "#173177"));
                 templateMessage.addData(new WxMpTemplateData("keyword3", LocalDateTime.now().format(df), "#173177"));
                 try {
-                    Thread.sleep(1000L);
+                    Thread.sleep(2000L);
                     wxMpService.getTemplateMsgService().sendTemplateMsg(templateMessage);
                 } catch (WxErrorException | InterruptedException e) {
                     result.add(grade);
